feature #866: separate navGraph
