+++ feature/fortune/src/main/java/org/sopt/official/feature/fortune/feature/fortundDetail/FortuneDetailScreen.kt
@@ -37,11 +37,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import kotlinx.serialization.Serializable
 import org.sopt.official.designsystem.SoptTheme
-
-@Serializable
-data class FortuneDetail(val date: String)
 
 @Composable
 fun FortuneDetailRoute(
